@@ -4,12 +4,12 @@ plugins {
 
 android {
     namespace 'com.example.doyinsave'
-    compileSdk 32
+    compileSdk 33
 
     defaultConfig {
         applicationId "com.example.doyinsave"
         minSdk 24
-        targetSdk 32
+        targetSdk 33
         versionCode 1
         versionName "1.0"
 
@@ -47,7 +47,15 @@ dependencies {
     implementation 'io.reactivex.rxjava3:rxandroid:3.0.2'
     implementation 'com.google.code.gson:gson:2.8.9'
     //
-    implementation 'com.google.android.gms:play-services-ads:21.3.0'
+    implementation 'com.google.android.gms:play-services-ads:22.5.0'
+    def billing_version = "6.0.1"
 
+    implementation "com.android.billingclient:billing:$billing_version"
 
+    implementation "androidx.lifecycle:lifecycle-extensions:2.2.0"
+
+    def lifecycle_version = '2.6.2'
+    implementation "androidx.lifecycle:lifecycle-process:$lifecycle_version"
+    implementation "androidx.lifecycle:lifecycle-runtime:$lifecycle_version"
+    annotationProcessor "androidx.lifecycle:lifecycle-compiler:$lifecycle_version"
 }
